+++ src/lib.rs
@@ -30,7 +30,7 @@
 //! Breaking changes ship as major version bumps; minor versions add modules
 //! or non-breaking helpers.
 
-#![doc(html_root_url = "https://docs.rs/doom-fish-utils/0.1.0")]
+#![doc(html_root_url = "https://docs.rs/doom-fish-utils/0.1.1")]
 #![cfg_attr(docsrs, feature(doc_cfg))]
 
 pub mod completion;